requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-haystack==2.5.1
elasticsearch==5.0.1
elasticsearch-dsl==5.0.0
gunicorn==19.6.0
psycopg2==2.6.2
python-dateutil==2.5.3
six==1.10.0
urllib3==1.19
wheel==0.26.0
whitenoise==3.2.2
